standardise code structure for the action rnn
